Basic Message Passing
Now we are beginning to take advantage of activeTab permission.
Remove popup.html to use the browser action to trigger a script instead
of opening up a popup.
Upon browser action click, run a script in the active tab. Script
captures click events and emits a message containing a property (src)
of the element.
Listen for message in extension, alert value as a test.