Add Browser Action Functionality
In injected script, check if the click event's target element is an
image. If so, send a message back and stop listening. Otherwise keep
waiting for a click on an image.
Move script to a separate file, ClickGetImg.js, as it is getting a bit
large for an inline function in text.
Refactor out a snip method so the context menu and browser action both
perform the same action with the targeted image.